Review comments fixed
