+++ src/Common/Dialogs/ConfirmationDialog.tsx
@@ -1,6 +1,8 @@
 import React from 'react'
 import { VisibleModal2 } from '../Modals/VisibleModal2'
-function ConfirmationDialog({ className = '', children }) {
+import { ConfirmationDialogBodyType, ConfirmationDialogButtonGroupType, ConfirmationDialogIconType, ConfirmationDialogType } from './Types'
+
+function ConfirmationDialog({ className = '', children }: ConfirmationDialogType) {
     return (
         <VisibleModal2 className="confirmation-dialog">
             <div className={`confirmation-dialog__body ${className}`}>{children}</div>
@@ -8,11 +10,11 @@ function ConfirmationDialog({ className = '', children }) {
     )
 }
 
-function Icon({ src, className = '' }) {
+function Icon({ src, className = '' }: ConfirmationDialogIconType) {
     return <img src={src} className={`confirmation-dialog__icon ${className}`} alt="" />
 }
 
-function Body({ title, subtitle = null, children = null }) {
+function Body({ title, subtitle = null, children = null }: ConfirmationDialogBodyType) {
     return (
         <div className="flex left column ">
             <h3 className="confirmation-dialog__title lh-1-25 dc__break-word w-100">{title}</h3>
@@ -22,7 +24,7 @@ function Body({ title, subtitle = null, children = null }) {
     )
 }
 
-function ButtonGroup({ children }) {
+function ButtonGroup({ children }: ConfirmationDialogButtonGroupType) {
     return <div className="flex right confirmation-dialog__button-group">{children}</div>
 }
