+++ src/Common/Dialogs/Types.ts
@@ -14,3 +14,19 @@ export interface ForceDeleteDialogType {
   forceDeleteDialogTitle: string
   forceDeleteDialogMessage: string
 }
+export interface ConfirmationDialogType {
+    className?: string
+    children: any
+}
+export interface ConfirmationDialogIconType {
+    src: string
+    className?: string
+}
+export interface ConfirmationDialogBodyType {
+    title: string
+    subtitle?: string
+    children?: any
+}
+export interface ConfirmationDialogButtonGroupType {
+    children: any
+}
