+++ src/Common/Error.tsx
@@ -1,10 +1,10 @@
 import React, { Component } from 'react'
 import EmptyState from './EmptyState/EmptyState'
 import notAuthorized from '../Assets/Img/ic-not-authorized.svg'
-import Reload from './Reload'
 import ErrorScreenNotFound from './ErrorScreenNotFound'
 import { ERROR_EMPTY_SCREEN } from './Constants'
 import GenericEmptyState from './EmptyState/GenericEmptyState'
+import Reload from './Reload'
 
 export class ErrorScreenManager extends Component<{
     code?: number
