+++ src/Common/Reload.tsx
@@ -1,17 +1,22 @@
 import React from 'react'
-import loadingFailure from '../Assets/Img/ic-loading-failure.png';
+import loadingFailure from '../Assets/Img/ic-loading-failure.png'
+import { ReloadType } from './Types'
 
-export default function Reload({ reload = null, className = '' }) {
-    function refresh(e){
-        window.location.reload();
+export default function Reload({ reload, className = '' }: ReloadType) {
+    function refresh(e) {
+        window.location.reload()
     }
     return (
-        <article className={`flex ${className}`} style={{ width: '100%', height: '100%' }}>
-            <div className="flex column" style={{ width: '250px', textAlign: 'center' }}>
-                <img src={loadingFailure} style={{ width: '100%', height: 'auto', marginBottom: '12px' }} alt="load-error" />
+        <article className={`flex w-100 h-100 ${className}`}>
+            <div className="flex column w-250 dc__align-center">
+                <img src={loadingFailure} style={{ height: 'auto' }} className="w-100 mb-12" alt="load-error" />
                 <h3 className="title dc__bold">Failed to load</h3>
-                <div className="dc__empty__subtitle" style={{ marginBottom: '20px' }}>We could not load this page. Please give us another try.</div>
-                <button type="button" className="cta ghosted" onClick={typeof reload === 'function' ? reload : refresh}>Retry</button>
+                <div className="dc__empty__subtitle mb-20">
+                    We could not load this page. Please give us another try.
+                </div>
+                <button type="button" className="cta ghosted" onClick={typeof reload === 'function' ? reload : refresh}>
+                    Retry
+                </button>
             </div>
         </article>
     )
